+++ keyboards/barbellboards/rollow/keymaps/t4corun/rules.mk
@@ -5,4 +5,5 @@
 #	keyboards\barbellboards\rollow\info.json
 #	keyboards\barbellboards\rollow\rules.mk
 
+LTO_ENABLE = no
 CONVERT_TO = promicro_rp2040
