+++ keyboards/klor/keymaps/t4corun/rules.mk
@@ -1,5 +1,7 @@
 # override userspace defaults
 
+ENCODER_MAP_ENABLE = yes
+
 # override keyboard defaults
 #	keyboards\klor\info.json
 #	keyboards\klor\rules.mk
@@ -7,6 +9,7 @@
 AUDIO_ENABLE = yes
 HAPTIC_ENABLE = yes
 RGB_MATRIX_ENABLE = yes
+SWAP_HANDS_ENABLE = no 
 
 # Choose your layout
 KLOR_CONFIG = saegewerk
