+++ qmk.json
@@ -2,7 +2,6 @@
     "userspace_version": "1.0",
     "build_targets": [
 
-		["ploopyco/trackball_nano", "lkbm"],
 		["crkbd/rev1", "t4corun"],
 		["bastardkb/charybdis/3x5/v2/splinky_3", "t4corun"],
 		["planck/rev6", "t4corun"],
