+++ users/t4corun/rules.mk
@@ -53,7 +53,11 @@ INTROSPECTION_KEYMAP_C += features/combo.c
 # include optional code for enabled features for each keyboard
 
 ifeq ($(strip $(OLED_ENABLE)), yes)
-  SRC += features/oled.c
+  ifeq ($(filter $(KEYBOARD), klor klor/2040),)
+    SRC += features/oled32.c
+  else
+    SRC += features/oled64.c
+  endif
 endif
 
 ifeq ($(strip $(AUDIO_ENABLE)), yes)
